Fixed Quick Sort, Added Heap Sort. Update Makefile to work for MinGW in Windows.

+++ Makefile
@@ -1,5 +1,5 @@
 CC = g++
-CCFLAGS = -lX11 -lGL -lpthread -lpng -lstdc++fs -std=c++17 -g
+CCFLAGS = -g -luser32 -lgdi32 -lopengl32 -lgdiplus -lShlwapi -ldwmapi -lstdc++fs -static -std=c++17
 
 all: array_main
 
@@ -25,4 +25,5 @@ commit:
 .PHONY: clean
 clean:
 		rm -f main *.o
-		rm array_main
+		[ ! -e array_main ] || rm array_main
+		[ ! -e array_main.exe ] || rm array_main.exe
